requirements.txt
tensorflow-recommenders==0.6.0
tensorflow-datasets==4.5.2
scann==1.2.5
pytest==6.2.3
pydantic==1.9.0
strictyaml==1.6.1
pandas-tfrecords==0.1.6
wget==3.2
pandas==1.4.2